Lot of instruments for difficulty - added LevelEvents in GameManager (Events which firing on certain levels) - added Dif's (special functions in gamemanager which affects bots HP and Damage on certain levels with certain periods) - Bonuses now disappear in some time - build update in ReadMe - Level equation was rewrite - Spawner now can change spawn chance of certain objects

+++ Assets/Script/Enemy/EnemyHealth.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class EnemyHealth : HealthPoints {
 
     [Tooltip("Needed only for enemy")]
-    public int PointsForDeath = 13;                 // how many points we will recive on enemy killing
+    public float PointsForDeath = 13;                 // how many points we will recive on enemy killing
 
     public bool DestroyAfterDeath = true;           // When owner is die, want you to destroy it, or just make invisible without collision
     [Tooltip("How many time will pass after death, before owner will be destroy (only if DestroyAfterDeath = true)")]
